feat: refactor Validator class and improve rule validation logic
- Refactored the Validator class to enhance validation rule handling and error messaging.
- Added a new method  to check for valid validation rules.
- Improved type safety with stricter type checks for validation rules.
- Enhanced error handling for invalid rules.
- Added detailed JSDoc comments for better code clarity and maintainability.
