+++ src/auth/index.ts
@@ -1,0 +1,4 @@
+
+export class Auth{
+    
+}
