+++ src/errors/index.ts
@@ -1,0 +1,4 @@
+
+export class Errors{
+    
+}
