+++ src/websocket/index.ts
@@ -1,0 +1,4 @@
+
+export class Websocket{
+    
+}
